Speed up drawing raw images
By changing the variable to 16 bit we can use an optimised buffer size
of 256 bytes that speeds up reading a 320x240 raw image  from SD Card
and plotting to screen in 405ms!

+++ TFT_ILI9341.h
@@ -252,7 +252,7 @@ class TFT_ILI9341 : public Print {
            pushColor(uint16_t color, uint16_t len),
 
            pushColors(uint16_t *data, uint8_t len),
-           pushColors(uint8_t  *data, uint8_t len),
+           pushColors(uint8_t  *data, uint16_t len),
 
            fillScreen(uint16_t color),
